+++ src/components/FullSizeConfirmModal/index.tsx
@@ -1,0 +1,5 @@
+export { default } from "./FullSizeConfirmModal";
+export type {
+  FullSizeConfirmModalProps,
+  FullSizeConfirmModalCloseReason,
+} from "./FullSizeConfirmModal";
