+++ src/components/FullSizeConfirmModal/styled.ts
