--- src/components/FullSizeModal/index.tsx
@@ -1,5 +1,0 @@
-export { default } from "./FullSizeModal";
-export type {
-  FullSizeModalProps,
-  FullSizeModalCloseReason,
-} from "./FullSizeModal";
