+++ src/constants/locale.ts
@@ -9,7 +9,7 @@ import {
 } from "..";
 import { SelectProps } from "../components";
 import { CreatableSelectProps } from "../components/CreatableSelect";
-import { FullSizeModalProps } from "../components/FullSizeModal";
+import { FullSizeConfirmModalProps } from "../components/FullSizeConfirmModal";
 import { EditFilterCardProps } from "../components/MultipleFilter/internal/EditFilterCard/EditFilterCard";
 import { FilterCardProps } from "../components/MultipleFilter/internal/FilterCard/FilterCard";
 
@@ -22,8 +22,11 @@ export interface Localization {
     ConfirmModal?: {
       defaultProps: Pick<ConfirmModalProps, "confirmText" | "cancelText">;
     };
-    FullSizeModal?: {
-      defaultProps: Pick<FullSizeModalProps, "confirmText" | "cancelText">;
+    FullSizeConfirmModal?: {
+      defaultProps: Pick<
+        FullSizeConfirmModalProps,
+        "confirmText" | "cancelText"
+      >;
     };
     FileUploader?: {
       defaultProps: Pick<FileUploaderProps, "title">;
@@ -83,7 +86,7 @@ export const jaJP: Localization = {
     ConfirmModal: {
       defaultProps: { confirmText: "確認", cancelText: "キャンセル" },
     },
-    FullSizeModal: {
+    FullSizeConfirmModal: {
       defaultProps: { confirmText: "確認", cancelText: "キャンセル" },
     },
     FileUploader: {
@@ -132,7 +135,7 @@ export const zhCN: Localization = {
     ConfirmModal: {
       defaultProps: { confirmText: "确认", cancelText: "取消" },
     },
-    FullSizeModal: {
+    FullSizeConfirmModal: {
       defaultProps: { confirmText: "确认", cancelText: "取消" },
     },
     FileUploader: {
